requirements.txt
openai
python-dotenv
rich
anthropic
